requirements.txt
# base utils
python-dotenv>=1.0
requests>=2.31
pydantic<2.12
urllib3<2.4

# LLM stacks
langchain==1.0.3
langchain-core==1.0.2
langchain-community==0.4.1
langchain-ollama==1.0.0
langgraph==1.0.2

# models and vector DB
transformers==4.57.1
sentence-transformers==5.1.2
chromadb==1.3.0
faiss-cpu==1.12.0

# data science
pandas==2.3.3
numpy==2.2.6
scikit-learn==1.7.2

# UI, serving, notebooks
streamlit==1.51.0
fastapi==0.120.2
uvicorn==0.38.0
gradio==5.49.1
jupyterlab==4.4.10

# torch for Apple Silicon
torch==2.9.0
torchvision==0.24.0
torchaudio==2.9.0

# dev tooling (optional but recommended)
black[jupyter]==25.9.0
flake8==7.3.0
pylint==4.0.2
isort==7.0.0

pytest==8.3.3